// HFY.MvcWebUI.Angular/ClientApp/src/app/lazy-general-page/classes/firmalar/models/firma-guncelle.ts
export class FirmaGuncelle {
  FirmaID: number
  Rol: string;
  BagliOlduguID: number;
  FirmaAd: string;
  FirmaKisaAd: string;
  Sifre:string;
  TelefonNumarasi: string;
  SMSTelefonNumarasi: string;
  GoogleMapsAdresi: string;
  VergiTCNo: string;
  FaturaAdresi: string;
  Adres: string;
  Eposta: string;
  PsiBar:string;
  DisDerinligiSayisi:number;
  ParaBirimID:number;
  KayitTarihi: Date;
  KullaniciGorevi:string;
  KullaniciKisaKod:string;
  YetkiliKisi:string;
  ListeAktiflik: boolean;
  FirmaHizmetler: number[];
  FirmaAracGruplar: number[];
}